modules/shaft/copy-requirements.txt
altgraph==0.17.4
black==24.10.0
bm3d==4.0.1
bm4d==4.2.3
certifi==2023.11.17
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
clihandler==0.6
colormath==3.0.0
colour-checker-detection @ git+https://github.com/colour-science/colour-checker-detection.git@a50651ba90603fbf4e1e1c216ceed4bb74dcc88f
colour-science==0.4.4
contourpy==1.2.1
cycler==0.12.1
defusedxml==0.7.1
distlib==0.3.9
et-xmlfile==1.1.0
filelock==3.16.1
fonttools==4.51.0
greenlet==3.0.3
identify==2.6.1
idna==3.6
imageio==2.34.1
isort==5.13.2
jira==3.5.2
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
lensfunpy==1.15.0
lxml==5.2.2
matplotlib==3.9.0
memory-profiler==0.61.0
mypy-extensions==1.0.0
networkx==3.3
nodeenv==1.9.1
numpy==1.26.4
oauthlib==3.2.2
opencv-python-headless==4.10.0.84
openpyxl==3.1.2
packaging==23.2
pandas==2.2.0
pathspec==0.12.1
pefile==2023.2.7
pillow==10.3.0
platformdirs==4.3.6
pre_commit==4.0.1
psutil==5.9.8
PyExifTool==0.5.6
pyfiglet==1.0.2
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.6
pyodbc==5.0.1
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
python-dateutil==2.8.2
pytz==2024.1
PyWavelets==1.6.0
pywin32-ctypes==0.2.2
PyYAML==6.0.2
QtAwesome==1.3.1
QtPy==2.4.1
rawpy==0.21.0
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.13.0
six==1.16.0
SQLAlchemy==1.4.51
threadpoolctl==3.5.0
tifffile==2024.9.20
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.1.0
virtualenv==20.27.0
statsmodels==0.14.4
